Created Multinomial classifier, cleaned up some...
Added Py to biasing for classification.  Changed input wordList - 1 to just wordList in problem 1 as it seemed that the index was already correct at the 0 position (it was giving the end of the wordList array when printed).  Multinomial might be able to be optimized... I couldn't find a way around looking at each individual wordcount in vocab array.  With fake data I produced the multinomial, looks to work fine 6/7 correct classifications when tested.  Mathematically, as far as I could tell with the fake data, our algorithm matches what we should be getting.  Going to start write up tomorrow morning.