Updated Kmeans
Added functionality to run Kmeans 10 times and pick the run that resulted in the lowest SSE total of all points (with respect to their cluster).  The kmeans function only performs kmeans once just in case we want to run kmeans individually for whatever reason..  The function kmeansTenTimes runs kmeans 10 times.  1-3 should be done, I can write the report when we finish up 4.